Update patch set 2
Patch Set 2:

The reason for this change is that most CPUs nowadays support half-precision floating-point numbers (or at least conversion between half- and single-precision). We don't use that functionality yet, but this change is in anticipation of it. It would save significantly on memory usage when any of these formats are used.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>