Update patch set 2
Patch Set 2:

This caused dEQP-GLES3.functional.shaders.struct.uniform.sampler_* to regress. I'm inclined to revert this change, because quite a bit of content that we know of uses samplers in structs, while apparently little content was affected by the uniform index going out of the sampler range.

I'm confused though because I think I recall that you mentioned samplers in structs to be one of the things that failed in your first attempt at fixing this but not your second which uses the map? Did an in-between version accidentally get uploaded and landed?

Let me know if you can spot the bug. I might have another look as well. If it's not a trivial fix I'll revert and we can have a deeper look at this later.

Patch-set: 2