WIP - Process independent vertex elements.
Previously, vertices would be processed in consecutive groups of four
(for SSE/NEON). Now four indices are read from the index buffer.
Reading the input was already a gather operation, but with constant
stride. The vertex cache now performs a scatter. The vertices are
written in reverse order so that the first vertex in a group is always
present in the cache.

Bug b/27351835

Change-Id: I19f2e47d90208dec07ce9ace8e9d520cf4a15cdd