Fix loops with dynamic iteration count.
The 'enable' execution mask for if/else, which is also used by loops,
was not being applied to predicate the instructions within a loop.

Also, we were resetting the execution mask for break statements to all
lanes at the end of a loop, but this is incorrect for nested loops and
unnecessary because we already inserted code to reset the mask to its
value before the loop.

Bug b/28710510

Change-Id: Ibd8e962ef60de76ec1c018ca064a791abf214c94