Update patch set 1
Patch Set 1: Code-Review-1

I really wonder if this couldn't be entirely handled within Program, since for every uniform and varying we have the full name of both the field and the struct. So there shouldn't be any need to pass that information from OutputASM a second time.

Also if I'm not mistaken we always declare all the member fields consecutively, so we can reasonably take advantage of that. So as soon as we encounter a struct that already has a location assigned to it, we can offset by the field index and the names and types should match.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1