Fix validation of statically referenced variables.
Variables such as varyings and in/out variables are verified to have
compatible types at link-time, and are considered active when
statically referenced anywhere in the code. This includes if/else
statements with a constexpr boolean condition (which is trivial to
optimize). So we still need to create AST nodes for these blocks, and
we still need to traverse them.

However, to optimize the compilation of 'uber-shaders', we can disable
emitting instructions for the trivially dead code paths.

Change-Id: I122b1cdcc4630c86a8ebfb4d4e37f3a7a335afbe