Update patch set 2
Patch Set 2: Verified-1

Fails dEQP-GLES3.functional.shaders.derivate.fwidth.static_if.float_highp and others. They have an if(false) which accesses a uniform, and and else block which accesses the same uniform. On the first visit we assign it a register, but don't declare it, while on the second visit the register is looked up but we don't declare it.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=-1