Allow GL_BGRA_EXT / GL_RGBA format combo.
GL_APPLE_texture_format_BGRA8888 specifies that when using the BGRA_EXT
format, it "must have an RGBA internal format". Note that
GL_EXT_texture_format_BGRA8888 also allows BGRA_EXT as internalformat.

Also support BGRA8 as a renderbuffer format. This is not specified in
the extensions, but is supported by all known implementations and is
assumed by Chromium.

Change-Id: I27e032054240efe96c32e27ebbdfb3c28213b1d8