Swap format/type in image upload.
This refactoring makes the code for texture image upload significantly
smaller. It also aligns better with the layout of table 3.2 in the
OpenGL ES 3.0 specification which lists valid format/type combinations.

Custom RGB565 format handling was removed because it corresponds with a
regular 2-byte per pixel memcpy.

Sized internal formats were removed because a recent change fixed not
using the internalformat parameter as upload format.

Change-Id: Ief6c6e60c3d9a734cac241022bc95ffc88d752c8