WIP - Fragment outputs
Fragment output arrays can be output to rendertargets of different types,
mixing integer and non integer types. For this purpose, we need a way to
track what type (Float, Int or UInt) is used inside the shader as the
fragment output in order to know if some conversion is required between
the shader's output type and the rendertarget's texture type.

Change-Id: I30e004449fb90713984b3481abb24c5d0cd6e867