Fix CopyTexImage internal format determination.
We were basing the determination of the effective sized internal format
of the image created by CopyTexImage on the combination of the
internalformat parameter (which can be unsized with OpenGL ES 2.0), and
the GL_UNSIGNED_BYTE type. Instead, if the internalformat parameter is
already sized we should use that directly instead.

If not sized, the spec contains complex rules for determining the
effective sized internal format. However, in practice only OpenGL ES
2.0 formats will have to be supported, and dEQP only tests with
GL_UNSIGNED_BYTE component types.

[OpenGL ES 3.05] section 3.8.5:
The error INVALID_OPERATION is generated if floating-point RGBA data is
required; if signed integer RGBA data is required and the format of the
current color buffer is not signed integer; if unsigned integer RGBA
data is required and the format of the current color buffer is not
unsigned integer; or if fixed-point RGBA data is required and the
format of the current color buffer is not fixed-point. The error
INVALID_OPERATION is also generated if the value of
FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING for the framebuffer attachment
corresponding to the read buffer is LINEAR (see section 6.1.13) and
internalformat is one of the sRGB formats described in section 3.8.16,
or if the value of FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING is SRGB and
internalformat is not one of the sRGB formats.

Note that floating-point formats are supported due to
GL_EXT_color_buffer_half_float.

Change-Id: I2e1b4f72449d874a5ca688819f98316a08ea86e8