Fix internal format handling for OpenGL ES 1.1.
With the egl::Image class now only storing the sized internal format,
some OpenGL ES 1.1 functionality broke because it was still expecting a
base format.

GL_NONE was replaced by GL_NO_ERROR when used as an error code because
it doesn't exist in GLES/gl.h. Note that both are defined as literal 0.

Also fix glBindRenderbufferOES(). It doesn't require anything except a
non-zero renderbuffer to create a new Renderbuffer object (it is not
required to use GenRenderbuffers before using BindRenderbuffer).
This had been fixed for GLES 3.0 and 2.0, but the same also applies
to GLES 1.1. Note that the Context already handles the case where
renderbuffer is zero properly.

Bug b/25983901

Change-Id: Ib41e91f371a06b2a09471ea45dd1f8f56b94091f