Fix renderbuffer binding for OpenGL ES 1.1.
glBindRenderbufferOES() doesn't require anything except a
non-zero renderbuffer to create a new Renderbuffer object (it is not
required to use GenRenderbuffers before using BindRenderbuffer).
This had been fixed for GLES 3.0 and 2.0, but the same also applies
to GLES 1.1. Note that the Context already properly handles the case
where renderbuffer is zero.

Bug b/25983901

Change-Id: If5e97be8027fca7c97cce5c44d063cd054443700