Check for an existing program before drawing
This adds a check for the program existing before any drawing takes
place. Otherwise, it's a no-op. I couldn't find anything in the spec
that says what's supposed to happen without a program bound other than
that it's UB and it's not an error:

OpenGL ES 3.0 spec 2.12:

"If UseProgram is called with program set to zero, then there is no
current program object, and the results of vertex and fragment shader
execution are undefined. However, this is not an error."

bug: 74108717

Previously, a draw call with no shader bound resulted in a segfault.
This CL just adds the check earlier, preventing the crash for the single
threaded case.

Test: Draw while program 0 is bound, in a single thread. No crash

Note that this does not fix the multithreaded case.
TODO: Don't crash when program in same share group is deleted while drawing

Change-Id: I31a7bc6ad5458107372542d20e51ea83d7503957