wip: multithreaded shader workflow
bug: 74108717

The multithreading bug when deleting a program and using it in a draw
call can fail in a few types of ways:

a. Context::useProgram and ResourceManager::deleteProgram interleave, so
we can end up setting the current program to an id of a program that is
null.

b. The entry point UseProgram (that calls Context::useProgram) has a few
chances to obtain different program objects under the same GLuint id,
making it so that if a program is deleted after the validation but
before the call to Context::useProgram, we will erroneously proceed into
Context::useProgram.

This CL:

1. Introduces a ScopedAtomic<T>, which is a safe way to access Resource
objects like Programs in a way that serializes all operations done on
the internal std::map data structure, and provides an extra scoped range
of safety when using the Resource object itself. To prevent value
passing and the unnecessary relocking, we also have
Context::ScopedAtomicNamespaceObject<T>.

2. Changes createProgram/deleteProgram/useProgram operations to take
place using ScopedAtomic<Program> objects in order to disallow
concurrent std::map operations, thus forcing concurrent Program binding
operations that affect std::map structure or refcount to be serialized.

3. Sets and gets the "current program" of the Context not as a GLuint id
but as an actual Program* pointer, so we don't end up retrieving it
multiple times during the course of operations and get different
Program* pointers for the same GLuint id, which happens a lot. Instead,
only set the current program while holding the lock over the std::map
name space.

Change-Id: I5842f483f4ecafe213a0052d8d3fdfdc7c970960