WIP-copyteximage-format
The OpenGL ES 3.0 spec section 3.8.5 states that:

If internalformat is unsized, the internal format of the new texel array is determined
by the following rules, applied in order. If an effective internal format exists
that has
1. the same component sizes as,
2. component sizes greater than or equal to, or
3. component sizes smaller than or equal to
those of the source buffer’s effective internal format (for all matching components
in internalformat), that format is chosen for the new image array, and this is also
the new texel array’s effective internal format.

Change-Id: Id243b963779108e205c275499ddfb6e041a873d6