Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> No assault intended :)
> 
> I see, that's a cool though different use case; I can't really properly evaluate the way cuttlefish does it, due to lack of context, but it seems fine to me to commit to Android-specific support in the EGL implementation if everything is designed to run on top of, I guess, Android graphics drivers.

Right now cuttlefish is rendering in the guest, but this is not a sustainable solution, because swiftshader's JIT is not allowed by Android sepolicy. These changes are only about moving the swiftshader instance onto the host side *without* expanding our use cases.

> Although if you removed _just_ the Android / EGL specificity, you'd greatly increase your range of viable GLES backends; like what some IHVs such as Intel/NVIDIA offer, plus ANGLE, or even the Android Emulator's translation library, without shader translation or mapping ES to desktop GL features (because that is what those GLES implementations often do underneath anyway). Something like a revamped gralloc + lightweight wrapper for ANativeWindow objects in the guest could go a long way toward that, and still let you keep the benefits of not translating.

I agree. The way I have modified the guest is to use drm_hwcomposer and minigbm from Chrome. I am using virtio-gpu with the virgl 3d acceleration enabled.

The difference is that I am trying not to mandate use of the virgl gallium driver or hardware 3d on either side -- this involves copying and translation, as well as pulling in mesa3d -- but that does not mean in the future we couldn't just enable that too, as a mode. Or, of course, use the android emulator's approach, but tunnel it through virtio-gpu.

Patch-set: 1
Reviewer: Gerrit User 6882 <6882@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>