Also look for SwiftShader in /system
This cleans up a problem that was introduced by
ag/Ic95a36f7108fcae8e6f120f0853ee00e9a8926c1

LOCAL_VENDOR_MODULE was first supported on O, so when we build
SwiftShader on earlier branches the libraries land in /system.
The fix in the Android.mk files would be complicated and would revert
some of the cleanup, and would cause issues for 64 bit only builds.
While it would be possible to extend the original approach, it just
seems cleaner to check in /system at runtime.

BUG: 76437145
Test: Local build and boot of N
Change-Id: I86e0363a8cbd7b8f2ba744f14a67dfe457f3725f