Also look for sibling libraries in /system
This cleans up a problem that was introduced by
https://swiftshader-review.googlesource.com/17188

LOCAL_VENDOR_MODULE was first supported on O, so when we build
SwiftShader on earlier branches the libraries land in /system.
The fix in the Android.mk files would be complicated and would revert
some of the cleanup, and would cause issues for 64 bit only builds.
While it would be possible to extend the original approach, it just
seems cleaner to check in /system at runtime.

Bug b/76437145
Test: Local build and boot of Cloud Android N
Change-Id: I86e0363a8cbd7b8f2ba744f14a67dfe457f3725f