Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

This looks OK to me, but note that the EGL spec does not make any guarantees about thread-local storage being deleted automatically when a thread exits. That's why it has eglReleaseThread. So strictly speaking this is a bug in the CTS. Drivers may provide automatic deletion of the TLS as a courtesy, but again, that's implementation-specific and should not be relied upon in portable code.

Could you file a CTS bug?

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2