Move MacOS EGL library copy logic from Chromium to SwiftShader
Chromium has multiple bundled apps on MacOS (Chromium and Content Shell).
Because of this, each bundled app needs a copy of the EGL libraries, so
in order to avoid having to have each bundled app make its own copy, the
copy targets will be moved into the libraries themselves (so Angle and
SwiftShader).

Bug chromium:757974

Change-Id: I0f2a67e46df91210be17a230b6c28f8af2e4e768