Don't flag temporary registers as used samplers.
This fixes a undefined shift if the temporary register has an index
higher than 32 (also, we only have 16 samplers currently). For GLSL
we declare all actual sampler units during ASM output, so we don't
have to rely on analyzing the instructions afterwards (that is only
still relevant for Direct3D 8/9 shaders).

Change-Id: I11a58964d53fcc2c29e0ad923d9a4a4161a545f4