Rectangle texture addressing fix
Texture rectangle coordinates were clamped in the [0, dim - 1] range,
but should have been sampled in the [0.5, dim - 0.5 range] according
to the spec (a related comment was added in the code).

Also, the 'fv' variable was getting overwritten within the address()
call for the 'w' parameter so another temporary variable, 'fz', was
added in order to avoid the possibility of 'fv' being overwritten.

Finally, initialized the z0 variable to 0 because its computation is
now skipped for rectangle textures and because the address() function
contains an early return that could potentially leave this variable
uninitialized.

Change-Id: I4bbc30b2a2b747eae2f2a1dfb710a986bff7849b