Fixed negative API regression
The format comparison for blitting from a multisampled buffer
was not ES2 specific, so the ES3 check was removed. That had
been added to fix blitting to the backbuffer on MacOS, which
is now instead fixed by allowing BGRA8 and RGBA8 to be
accepted interchangeably as draw and read formats.

Change-Id: Id577372791007a780c542b0986378147e1e7bc1b