Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Sergey, I'm not sure where things stand with using SwiftShader on Fuchsia, but we're trying to eliminate the use of OSMesa and Alexis found this change to be necessary. Would you mind having a look?

Currently we use OSMesa on Fuchsia just for layout tests. But the layout tests are not enabled on any bots at the moment. It should be possible to switch them to SwiftShader, but it's questionable if it's useful because long term we are planning to support only Vulkan for rendering. Feel free to break/disable SwiftShader and OSMesa on Fuchsia if it blocks anything.

Patch-set: 1
Reviewer: Gerrit User 6844 <6844@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>