Fix protecting exported symbols from overrides.
While all symbols except for the ones in the version script are hidden,
the exported ones can still be overridden by other libraries which
export the same symbols. Using -fvisibility=protected fixes this. We can
only apply it to the shared library targets because of ld linker issues.

Bug b/110884149

Change-Id: I98222420eabbb9ea0873816f52e1354625b9a3a8