Protect exported symbols from overrides.
While all symbols except for the ones in the version script are hidden,
the exported ones can still be overridden by other libraries which
export the same symbols. Using -fvisibility=protected or the equivalent
attribute fixes this.

Protected visibility used to not work because of
https://bugs.llvm.org//show_bug.cgi?id=30960, but that doesn't appear
to be an issue any more.

Bug chromium:852537
Bug swiftshader:64

Change-Id: Id0b1197b90baa5f5c68bf2aa107d12f8a9856796