Guard against out of bounds accesses caused by a large base level
A texture's base level can be any positive integer, but SwiftShader
stores texture levels in a array that has an implementation
dependent size (IMPLEMENTATION_MAX_TEXTURE_LEVELS). To avoid
accessing this array out of bounds, a check was added at the
beginning of any function using the base level to index the array.

Change-Id: I9b439018f209a47371bd2959ba847345326964dd