Fix texture completeness test when using sampler objects.
When a sampler object is bound to a texture unit, it overrides the
texture's filter modes. A texture without mipmaps can thus be complete
even if it has a filter mode which requires mipmaps, if there's a
sampler object with a filter mode which doesn't require mipmaps.

Fixes https://github.com/google/filament/pull/220

Change-Id: I200c0bba467d904ac0fd6d93b935e052c47b2030