Update patch set 1
Patch Set 1:

+Nicolas and Alexis for vetting this CL

If this is really all that's needed to get SwiftShader building on 32-bit ARM then it would be great to support this. Nicolas did some investigation into this previously and found that at least for Android there were missing NDK headers which would make this more difficult.

Patch-set: 1
CC: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
CC: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
CC: Gerrit User 5220 <5220@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>