Update patch set 5
Patch Set 5:

This works as an ICD now. I added some instructions in the description on how to use it.

The library path in swiftshader_vk.json is a bit unconventional for Windows, but it works.
Specifying a path like "D:\..." doesn't work in this context. Not sure why exactly.

Patch-set: 5
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>