Update patch set 2
Patch Set 2:

If SSE 4.1 is enabled, `@llvm.floor.f32(<4 x float>)` can be converted to `roundss`.  (p.s. SSE3 is not sufficient.)

If NEON is enabled for AArch64, it can be converted to `frintm`.

If NEON is enabled for ARMv8 (32-bit), it can be converted to `vrintm.f32`.  (p.s. armv7 is not sufficient.)

Here are the test code:

    declare float @llvm.floor.f32(float %b)

    define void @foo(float* %out_ptr, float *%in_ptr) {
    ent:
      %in = load float, float *%in_ptr
      %out = call float @llvm.floor.f32(float %in)
      store float %out, float* %out_ptr
      ret void
    }

And I tested with the following commands:

    llc test.ll -mtriple x86_64-linux-gnu -mattr=+sse4.1
    llc test.ll -mtriple aarch64-linux-gnu -mattr=+neon
    llc test.ll -mtriple armv8-linux-gnueabihf -mattr=+neon

    llc test.ll -mtriple x86_64-linux-gnu -mattr=+sse3  // XX
    llc test.ll -mtriple armv7-linux-gnueabihf -mattr=+neon  // XX

Patch-set: 2
Reviewer: Gerrit User 7038 <7038@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>