Replace all-makefiles-under makefile-macro with explicitly including Android.mk(s) from subdirectories that are related to OpenGL and are only needed by Android builds.
The change are mainly made to the top-level Android.mk and the
src/Android.mk. The other three sub-Android.mk(s) are deprecated and
deleted. Soong compile-time is expected to shorten with no use of
all-makefiles-under macros.

Since the project co-existed with both Android.bp and Android.mk.
The test was performed in absence of Android.bp on oc-mr1-dev branch
against aosp_arm-eng and aosp_arm64-eng targets.

Bug: b/29023322
Test: oc-mr1-dev branch, aosp_arm-eng (LLVM3) & aosp_arm64-eng (LLVM7) targets
Change-Id: I05ac46213f849747621c7084bca2549a22513881