Update patch set 4
Patch Set 4:

We've already moved away from LLVM 3.0 due to b/116540140 The TLDR is that LLVM 3.0 fails criteria #2 below, and subzero did not, so we made an emergency switch. That's in part because the memory leak in SwiftShader leads to destabilization of the Android presubmit system, and that probably led to requiem/doc/postmortem101566

It seems to me that a move to LLVM 7.0 would be good to have, but doesn't qualify as an emergency. I'd propose that we test, applying the same criteria that we did for the emergency switch:

1. A full CTS run with ASAN with SwiftShader crashes due to ASAN failures. subzero is able to do this, and heap corruption is just as dangerous as a memory leak.
2. No obvious leaking with the test that we used to validate subzero before we did the emergency switch.
3. No serious regressions in the graphics results for CTS (both deqp and the CTS tests) relative to subzero.

We can get to that by merging onto the cuttlefish-testing branch of AOSP. Merging there is safe since it doesn't directly affect master. We're equipped to run the CTS and leak checks.

If those look promising I'd be in support of a merge to master, but not before.

Patch-set: 4
Reviewer: Gerrit User 5200 <5200@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>