Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> We've already moved away from LLVM 3.0 due to b/116540140

This review is for the upstream SwiftShader repository's master branch, which still defaults to LLVM 3.0.

> It seems to me that a move to LLVM 7.0 would be good to have, but doesn't qualify as an emergency.

Development changes shouldn't have to qualify as an emergency to be approved.

If SwiftShader had no JIT-compiler, and you'd have to choose between enabling LLVM 3.0 or LLVM 7.0, which one would you pick? It seems an easy one to me considering that LLVM 3.0 is known to be flawed and LLVM 7.0 is the latest-and-greatest version which we know to be heavily tested and also didn't exhibit any glaring leakage in my smoke test.

Even in the unlikely case that LLVM 7.0 is somehow worse than LLVM 3.0, we still want to deprecate the latter and fix 7.0 so we can move forward. It seems better to do that now than to take the risk during an emergency.

Patch-set: 4
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>