Update patch set 4
Patch Set 4:

LLVM 3.0 isn't a factor for us on Android. We use subzero, so we're not blocking deprecation of LLVM 3.0 at all.

I didn't realize that this was on the SwiftShader repository. Sorry for blocking here. We can reapply the subzero change when we update on the Android repositories.

I apologize for the confusion.

Patch-set: 4
Reviewer: Gerrit User 5200 <5200@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=0