Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> (1 comment)
> 
> Thanks! I assume this passes ReactorUnitTests?
> 
> Are you planning to also have an optimized version that uses the VSHF instruction? SwiftShader makes very heavy use of shuffles so this could impact performance a lot.

Sorry for the late reply.

Sadly the VSHF instruction cannot be used as it belongs to the MIPS SIMD Architecture set of instructions which is not supported in Subzero.

Since there are tests that still failing (unittests & SubzeroTest)you may want to hold off committing this for now.
The tests are failing either because of the introduced change or because there were other errors all along, but are showing since the implementation in question been supplied so the code paths that follow can now be taken.
I'm looking into this.

I needed to update the SwiftShader repository to have the ReactorUnitTests make rule available, I'll give an update when the build gets finished.
It may take some time, building natively takes rather long with the hardware I'm using.

Patch-set: 1
Reviewer: Gerrit User 7076 <7076@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>