Fixed warnings in the Vulkan project
Some classes were given the following attribute:
[[clang::lto_visibility_public]]

This produces a warning in Visual Studio.

This was meant to allow these base classes to be visible
to both libEGL and libGLESv2. Vulkan will be a single DLL,
so this won't be an issue, so I simply removed both
instances of this attribute in the Vulkan project.

Change-Id: I6d1b0bb0ab5351d6b788917c0b876e9355505404