Update patch set 4
Patch Set 4:

(6 comments)

I have concerns that going directly to a triangle is a big leap. Instead we could focus on doing on "GPU memcpy" with vkCmdCopyBuffers. Also I don't think the allocator is more complex than the boilerplate needed for the inheritance scheme, but that might be subjective.

Patch-set: 4