Update patch set 2
Patch Set 2:

This is how 100% of VkPromotedExtensions.cpp was implemented, so it needs to be fixed too. We probably need common functions that can be called by either regular or extension function for all of these cases.
Also, we originally had vk::... functions for all of these, but they were removed because it was decided that we'd use the API functions directly. Should we have kept the vk::... functions for these cases? Seems better than duplicating code.

Patch-set: 2
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>