Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Issue that we've seen on mips64el Chromium cross-compile build [1] and the issue reported for aarch64 [2], appear when SwiftShader is built for clang_x64_v8_mips64el or clang_x64_v8_arm64. Even when cros-compiling, there are parts built for the host,
> so in this case LLVM_NATIVE_TARGETINFO is correctly defined as LLVMInitializeX86TargetInfo in [3], so swiftshader_llvm_x86 sources are indeed needed for cross-compile builds.
> 
> [1] http://www.rt-rk.com/mips-buildbot/builders/build_and_test_chromium_trunk_for_MIPS64/builds/240/steps/Build%20Chromium/logs/stdio
> [2] https://bugs.chromium.org/p/chromium/issues/detail?id=905480
> [3] https://cs.chromium.org/chromium/src/third_party/swiftshader/third_party/llvm-7.0/configs/linux/include/llvm/Config/llvm-config.h?gsn=HistogramSamples&q=llvm-7.0/configs/linux/include/llvm/Config/llvm-config.h&g=0&l=134

Thanks for taking another look! I wonder if we can get rid of that host dependency. I know LLVM builds some executables which then generate files which are used in the rest of the build. But in our case we've pre-generated those and placed them in configs/common.

Anyway, might be more effort than what it's worth to avoid that host dependency.

Patch-set: 2