Update patch set 1
Patch Set 1: Verified-1

Unlike some of the other pipeline state related changes, this one is not entirely trivial. So I'd prefer not to land this until we have a triangle on the screen and can verify that this doesn't break anything. Hence the Verified-1. We know the current code works, so let's just translate/hardcode the one topology mode we need for the first triangle.

Patch-set: 1
CC: Gerrit User 5065 <5065@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
CC: Gerrit User 6982 <6982@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=-1