Update patch set 1
Patch Set 1: Code-Review-1

I don't think this change is correct. See the EGL spec for details about how this should work.

https://www.khronos.org/registry/EGL/specs/eglspec.1.5.pdf

"When the current rendering API is EGL_OPENGL_ES_API or EGL_OPENGL_-
API, the values of EGL_CONTEXT_MAJOR_VERSION18 (the major version) and
EGL_CONTEXT_MINOR_VERSION (the minor version) request creation of an
OpenGL ES or OpenGL context, respectively, supporting the specified version
major.minor of that client API. The context returned must be the specified version, or a later version which is backwards compatible with that version. Even if a
later version is returned, the specified version must correspond to a defined version
of the client API."

IOW, if you request 2, you might get 3.. and..

"Typically, the implementation will return the most recent version of OpenGL it
supports which is backwards compatible with the requested version."

Which is what SS is now doing.. and..

"Querying the GL_VERSION string with glGetString in either OpenGL or
OpenGL ES (or the GL_MAJOR_VERSION and GL_MINOR_VERSION values with
glGetIntegerv, in an OpenGL 3.0 or later context) will return the actual version
supported by a context."

I'm reading "actual version" here as the version it substituted for the request (so if you requested 2 and got 3, it'll return 3), but the language does leave some room for interpretation.

Patch-set: 1
Reviewer: Gerrit User 6882 <6882@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1