Update patch set 1
Patch Set 1:

This is incorrect. The spec says:

`float32 are the color clear values when the format of the image or attachment is one of the formats in the Interpretation of Numeric Format table other than signed integer (SINT) or unsigned integer (UINT). Floating point values are automatically converted to the format of the image, with the clear value being treated as linear if the image is sRGB.`

Patch-set: 1
Reviewer: Gerrit User 6982 <6982@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>