Support for clearing Framebuffer attachments
This is the implementation for vkCmdClearAttachments.

It includes the new command along with the plumbing to
get the information to the Image clear function.

The only non trivial behavior here is that the provided
baseArrayLayer for the clear command is to be offset by
the ImageView's subresourceRange's baseArrayLayer so
that it clears the correct layers.

Bug b/119621736

Passes all tests in:
api.image_clearing.dedicated_allocation.clear_color_attachment.*
api.image_clearing.dedicated_allocation.clear_depth_stencil_attachment.*

Change-Id: I19f86b63239ca2fb4d53f2c6c9cb75185794f061