Update patch set 3
Patch Set 3:

This is a C++11 vs C++14 thing -- whether std::hash<enum type> exists.

The cmake-based linux build has been building as C++14 for a while due to some things that crept in (make_unique etc), and the windows build is apparently building as "whatever that version of VS wants to do".

Patch-set: 3