Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> lfy@ etc -- Great that you identified what was missing, but let's not forget the other side of making this right;
> 
> Some possibilities:
>   (1) We clearly document somewhere that ES3 support on Android implies that OES_EGL_image_external_essl3 is supported (and complain loudly in CTS if that's violated), or
>   (2) Skia needs to learn to cope with its absence (presumably by dropping to ES2 shaders), or
>   (3) Skia needs to give up loudly on such implementations, such that we wouldn't have spent a huge amount of time trying to figure out what was broken.

We discussed this briefly in bug 78110728 and it did end up in the CDD, but yeah there is no test for that :( Skia uses essl3 together with es3 unconditionally afaik. CtsSkqpTestCases will definitely fail a lot if this is missing but the reason for missing is not spelled out, iirc. How about: this should be something that is loudly signaled even outside of CTS tests; once the device boots and tries to use hwui skia engine without essl3 ext, there should be log spam.

Patch-set: 5