Add Visual Studio project files for SPIRV-Tools.
SPIRV-Tools comes with CMake build files. Integrating it into our
custom Visual Studio solution requires making the project files that
it generates path independent so that they can be checked into the
repository.

This is achieved by the build/cmake.sh shell script that is intended
to be run from the Git Bash shell. It needs to be re-run when any
CMakeLists.txt file changes.

.gitignore was modified to allow for build files under the build/
directory to be checked in. CMakeFiles/ and CMakeCache.txt are ignored
because they contain absolute paths and aren't necessary for building.

Bug b/123642959

Change-Id: I0706e3302d16144ecb9c0802cec897aa66affe6f