Update patch set 2
Patch Set 2:

I notice this isn't sent out for review yet, but, this looks awesome! Please file a tracking bug at https://g.co/swiftshaderbugs (keep in mind these are public, there's also a private Buganizer component if you need it) to centralize the discussion around this and be able to relate the CLs.

Patch-set: 2
CC: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>