CommandPool reset and trim
Implemented vkResetCommandPool and vkTrimCommandPool.
The reset command out all command buffers in the command
pool back to the initial state. It also frees up the entire pool.

The trim command is a potential future memory usage optimization.
It is currently a noop.

Passes a few tests in dEQP-VK.api.command_buffers.

Bug b/119827933

Change-Id: Ic974f63a8dfca014462ac904218b9dcc3035fc8a