Update patch set 1
Patch Set 1:

FYI, you can push to refs/for/master%wip instead of refs/for/master to mark a change as work-in-progress immediately, which will avoid triggering sending out an e-mail.

Anyway, while I'm peeking here: let's land https://swiftshader-review.googlesource.com/c/SwiftShader/+/24648 first, which is a small subset of this and would be a smaller revert if clusterfuzz detected it as an issue.

Patch-set: 1
CC: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Work-in-progress: false