GLES: Fix race in Renderer destruction
The clipper and blitter were being destructed before the worker threads were stopped.

Move the call to terminateThreads() to the top of the destructor to fix this.

Also nullify after deleting other fields - this would have been much harder to detect if the clipper and blitter were not already being nullified after destruction.

Bug: b/124288179
Change-Id: Iefea829dfb8556c877c0a086253c5924c54aa638