Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This is awesome, thanks for working on this! Just a few very early remarks:
> - This will also need an implementation in SubzeroReactor.cpp, and one for LLVM 3.0. Currently it's a build option to select between LLVM 7.0, LLVM 3.0, or Subzero. I'll make a note about testing all of them with Kokoro like we used to with Travis CI. Anyway, I'm fine with the LLVM 3.0 and Subzero implementations being empty for now.
> - Doesn't need to happen in this patch, but I was wondering if we could generalize calling external functions a bit, i.e. let users of Reactor pass in the function and let template magic (similar to Function<>) figure out the signature so we can call it in a type-safe manner. Obviously functions with a variadic argument like printf are more of a challenge and a custom Printf() implementation is totally fine.
> - I'm a bit apprehensive about adding a function to extract bytes from Byte4 vectors, at least to public interface. Part of Reactor's design philosophy is to only add operators that are efficient to implement on the majority of CPUs. This is why we have long repetitive implementations of operators for each type, with some of them missing here and there. It helps to implicitly write efficient code (code that takes multiple instructions requires multiple Reactor operators or intrinsics). That said, for Printf we obviously don't care about performance to that extent and we have any internal utility function to help us out.

Glad you like it - printf() is the Google Way of Debugging™. :-)

Note that this is pure debug code - I do not expect anyone to be printf'ing from JIT'd code in production. With that said:

> This will also need an implementation in SubzeroReactor.cpp,

I'd argue we don't for now. If you attempted to use this with the subzero backend, you'll get a linker error - which is a pretty clear indication that it isn't supported by that backend (yet).

> I was wondering if we could generalize calling external functions a bit

I agree this would be useful, especially for syncing compute barriers, and the like as we discussed. However, I'd prefer to do this generalisation when there are functions to call in production code.

> I'm a bit apprehensive about adding a function to extract bytes from Byte4 vectors

A fair point. I'm not entirely sure how I'd cleanly hide this away, unless I wrap all of this in a DEBUG preprocessor. I'll have a think.

Note, this is far from being ready for review or submitting. I'm just using it to figure out what's going on with these cubemap crashes.

Patch-set: 2
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>