GLES: Fix for crashes when sampling certain textures.
There was an optimization that skipped a [0, 1] clamp for a particular sampling mode as it assumed the input coordinates were always within these bounds.

However, if the texture sample coordinates were inf or NaN, this assumption broke, causing the returned address to be  outside the bounds of the image data.

Change-Id: I0af34ee4c2792b19081d9270fd0b1e0d0559287e