Add regres tool.
Regress is a tool that detects test regressions with SwiftShader changes.

Regres monitors changes that have been put up for review with Gerrit.
Once a new patchset has been found, regres will checkout, build and test the change against the parent changelist. Any differences in results are reported as a review comment on the change.

Change-Id: Iec77d56bb2fc79153c090226e668e613b7969bff