Update patch set 2
Patch Set 2:

(1 comment)

I'm apprehensive about this one. Attributes are, besides the standardized ones, compiler implementation specific. They *must* be ignored and not cause an error by compilers that don't understand them. Hence treating any warning about unknown attributes as error is violating that, and any workaround to avoid the error seems like fighting the wrong issue. Attributes were intended to avoid having to use preprocessor macros for writing portable code.

Specifically for this case, it's not clear that LTO_VISIBILITY_PUBLIC is even an attribute without looking up its definition. Hence in my opinion it is hurting readability (renaming it to CLANG_LTO_VISIBILITY_PUBLIC_ATTRIBUTE wouldn't be much better in that regard). Another issue is that in general it's hard to enforce the use of a macro like this in future refactorings. Even with warnings-as-errors enabled it can lead to redefining the same macro elsewhere. Granted, it's probably not an issue in this legacy code, but it doesn't set a good precedent.

Anyway, I don't mean to be contrarian if this is considered valuable to others. Also it's in fairly isolated code that will be deprecated at some point so I don't care terribly.

That said, I can't think of a significant disadvantage of just suppressing this warning. If an attribute was misspelled it's not going to have the desired effect on the compiler you're working on, which should be hard to miss.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>